+++ crates/tests/implement_interface/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "test_interface"
+name = "test_implement_interface"
 version = "0.0.0"
 authors = ["Microsoft"]
 edition = "2018"
